requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
mysql-connector-python
openai